Major update

+++ Backend/flaskrestaurantapp/routes.py
@@ -1,10 +1,6 @@
 from flaskrestaurantapp.models import User, Products
-from flask import Flask, session
-import flask
 from flaskrestaurantapp import app
 from flask import jsonify, request
-import json
-import functools
 from flaskrestaurantapp import db
 from flask_jwt_extended import create_access_token
 from flask_jwt_extended import jwt_required
@@ -12,10 +8,15 @@ from flask_jwt_extended import set_access_cookies
 from flask_jwt_extended import unset_jwt_cookies
 from flask_jwt_extended import create_refresh_token
 from flask_jwt_extended import get_jwt_identity
+from flask_jwt_extended import get_jwt
 from flask_bcrypt import Bcrypt
+from datetime import timedelta, timezone, datetime
+import json
+import flask
 
 
 bcrypt = Bcrypt(app)
+
 
 # Authentication
 # =============================== 
@@ -28,21 +29,27 @@ def register():
     email = data['email']
     hashed_password = bcrypt.generate_password_hash(data['password']).decode('utf-8')
     if not username:
-        return jsonify({"msg": "Missing username parameter"}), 400
+        message = jsonify({"msg": "Missing username parameter"})
+        return message, 400
     if not email:
-        return jsonify({"msg": "Missing email parameter"}), 400
+        message = jsonify({"msg": "Missing email parameter"})
+        return message, 400
     if not hashed_password:
-        return jsonify({"msg": "Missing password parameter"}), 400
+        message = jsonify({"msg": "Missing password parameter"})
+        return message, 400
     user = User.query.filter_by(username=username).first()
     if user:
-        return jsonify({"msg": "Username already exists"}), 400
+        message = jsonify({"msg": "Username already exists"})
+        return message, 400
     user = User.query.filter_by(email=email).first()
     if user:
-        return jsonify({"msg": "Email already exists"}), 400
+        message = jsonify({"msg": "Email already exists"})
+        return message, 400
     user = User(username=username, email=email, password=hashed_password)
     db.session.add(user)
     db.session.commit()
-    return jsonify({"msg": "User created successfully"}), 200
+    response = jsonify({"msg": "User created successfully"})
+    return response, 200
   
 
 @app.route('/login', methods=['POST'])
@@ -61,15 +68,33 @@ def login():
         return jsonify({"msg": "Bad username or password"}), 401
     if not bcrypt.check_password_hash(user.password, password):
         return jsonify({"msg": "Bad username or password"}), 401
-    access_token = create_access_token(identity=user.role)
-    refresh_token = create_refresh_token(identity=user.role)
-    resp = jsonify({'message': 'Login successful'})
-    set_access_cookies(resp, access_token)
-    resp.set_cookie('refresh_token', refresh_token, httponly=True)
-    return resp, 200
+    access_token = create_access_token(identity=user.email)
+    refresh_token = create_refresh_token(identity=user.email)
+    response = jsonify({'message': 'Login successful'})
+    set_access_cookies(response, access_token)
+    response.set_cookie('refresh_token', refresh_token, httponly=True)
+    return response, 200
+
+
+@app.after_request
+def refresh_expiring_jwts(response):
+    try:
+        exp_timestamp = get_jwt()["exp"]
+        now = datetime.now(timezone.utc)
+        target_timestamp = datetime.timestamp(now + timedelta(minutes=30))
+        if target_timestamp > exp_timestamp:
+            access_token = create_access_token(identity=get_jwt_identity())
+            data = response.get_json()
+            if type(data) is dict:
+                data["access_token"] = access_token 
+                response.data = json.dumps(data)
+        return response
+    except (RuntimeError, KeyError):
+        return response
+
  
   
-@app.route('/logout', methods=['POST'])
+@app.route('/logout', methods=['DELETE'])
 def logout():
     response = jsonify({'message': 'Logout successful'})
     unset_jwt_cookies(response)
@@ -105,32 +130,30 @@ def getProducts():
     output = []
     for Products in productsList:
         currentProduct = {}
+        currentProduct['id'] = Products.id
         currentProduct['product_name'] = Products.product_name
         currentProduct['price'] = Products.price 
         currentProduct['category'] = Products.category
-        currentProduct['description'] = Products.description 
+        currentProduct['main_category'] = Products.main_category
         currentProduct['ingredients'] = Products.ingredients
-        currentProduct['allergens'] = Products.allergens
         output.append(currentProduct)
     return jsonify(output), 200
 
 
 @app.route('/products/<int:id>/', methods=['PUT'])
 @jwt_required()
-def protected():
+def updateProducts(id):
     current_identity = get_jwt_identity()
-    if current_identity == 'admin' or current_identity == 'super':
-        def updateProducts(id):
-            data = request.get_json()
-            currentProduct = Products.query.get(id)
-            currentProduct.product_name = data['product_name']
-            currentProduct.price = data['price']
-            currentProduct.category = data['category']
-            currentProduct.description = data['description']
-            currentProduct.ingredients = data['ingredients']
-            currentProduct.allergens = data['allergens']
-            db.session.commit()
-            return jsonify({'message': 'Product updated!'}), 200
+    if current_identity == 'admin@yahoo.com' or current_identity == 'super@yahoo.com':
+        data = request.get_json()
+        currentProduct = Products.query.get(id)
+        currentProduct.product_name = data['product_name']
+        currentProduct.price = data['price']
+        currentProduct.category = data['category']
+        currentProduct.main_category = data['main_category']
+        currentProduct.ingredients = data['ingredients']
+        db.session.commit()
+        return jsonify({'message': 'Product updated!'}), 200
     else:
         return jsonify({'message': 'Acces forbidden!'}), 401
 
@@ -139,26 +162,33 @@ def protected():
 @app.route('/products/<int:id>/', methods=['DELETE'])
 @jwt_required()
 def  deleteProduct(id):
-    currentProduct = Products.query.get(id)
-    db.session.delete(currentProduct)
-    db.session.commit()
-    return jsonify({'message': 'Product deleted!'}), 204
+    current_identity = get_jwt_identity()
+    if current_identity == 'admin@yahoo.com':
+        currentProduct = Products.query.get(id)
+        db.session.delete(currentProduct)
+        db.session.commit()
+        return jsonify({'message': 'Product deleted!'}), 204
+    else:
+        return jsonify({'message': 'Acces forbidden!'}), 401
 
 
 @app.route('/products', methods=['POST'])
 @jwt_required()
 def postProducts():
-    productData = request.get_json()
-    product = Products(
-                product_name=productData['product_name'], 
-                price=productData['price'], 
-                category=productData['category'], 
-                description=productData['description'], 
-                ingredients=productData['ingredients'], 
-                allergens=productData['allergens']    
-            )
-    db.session.add(product)
-    db.session.commit()
-    return jsonify(productData), 201
+    current_identity = get_jwt_identity()
+    if current_identity == 'admin@yahoo.com' or current_identity == 'super@yahoo.com':
+        productData = request.get_json()
+        product = Products(
+                    product_name=productData['product_name'], 
+                    price=productData['price'], 
+                    category=productData['category'],
+                    ingredients=productData['ingredients'],
+                    main_category=productData['main_category']     
+                )
+        db.session.add(product)
+        db.session.commit()
+        return jsonify(productData)
+    else:
+        return jsonify({'message': 'Acces forbidden!'}), 401
 
 # ===============================
